Refactor userspace code
remove `process.go` file and condense all
userspace functionality into `ebpf.go`
and `types.go`

Fix some vendoring problems

Bump some k8s dependencies

smooth out any vendoring issues
