+++ backends/windows/userspace/go.mod
@@ -4,8 +4,8 @@ go 1.18
 
 require (
 	github.com/spf13/pflag v1.0.5
-	k8s.io/api v0.24.1
-	k8s.io/apimachinery v0.24.1
+	k8s.io/api v0.24.2
+	k8s.io/apimachinery v0.24.2
 	k8s.io/klog/v2 v2.60.1
 	k8s.io/utils v0.0.0-20220210201930-3a6ce19ff2f9
 	sigs.k8s.io/kpng/api v0.0.0-20220521134046-f747cedbe766
